requirements.txt
amqp==5.1.1
aniso8601==9.0.1
appdirs==1.4.4
argcomplete==3.2.2
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.29.0
attrs==23.2.0
billiard==4.1.0
botocore==1.34.64
celery==5.3.1
certifi==2023.5.7
cfgv==3.4.0
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.4
cron-descriptor==1.4.0
distlib==0.3.8
Django==4.2.3
django-admin-thumbnails==0.2.8
django-appconf==1.0.6
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-imagekit==5.0.0
django-js-asset==2.2.0
django-modeltranslation==0.18.10
django-pgtrigger==4.11.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
Faker==19.2.0
filelock==3.13.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.5.35
idna==3.4
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.1
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.5
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.8.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.12.1
pilkit==3.0
Pillow==10.0.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.2.0
polib==1.2.0
promise==2.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pypng==0.20220715.0
pytest==7.4.0
pytest-django==4.5.2
python-barcode==0.15.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-escpos==3.1
python-monkey-business==1.0.0
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0
qrcode==7.4.2
redis==4.6.0
referencing==0.33.0
regex==2024.7.24
reportlab==4.0.9
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.0.3
usb==0.0.83.dev0
uvicorn==0.22.0
vine==5.0.0
virtualenv==20.25.1
wcwidth==0.2.6
websockets==11.0.3
zope.event==5.0
zope.interface==6.2
